* move pytorch dataset to new location and enable tests
