--- data/pytorch/BUILD
@@ -1,14 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-py_library(
-    name = "torch_dataset",
-    srcs = ["torch_dataset.py"],
-    deps = ["//recordio:lib"]
-)
-
-# TODO: need to add torch package to docker image.
-#py_test(
-#    name = "torch_dataset_test",
-#    srcs = ["torch_dataset_test.py"],
-#    deps = [":torch_dataset"],
-#)
